Changed CMAKE_RUNTIME_OUTPUT_DIRECTORY

+++ cmake_Unix.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 
 cd build
-rm -R CMakeFiles
-rm Makefile
+
+# Удаление устаревших файлов
+if [ -d CMakeFiles ]; then
+    rm -R CMakeFiles
+fi
+
+if [ -f Makefile ]; then
+    rm Makefile
+fi
+
 ./cmake.sh
